+++ addons/eventsheet/editor/dock/dock_input_dispatch.gd
@@ -70,6 +70,8 @@ func on_row_context_menu_id_pressed(id: int) -> void:
 			_dock._copy_selection_as_text()
 		_dock.ROW_MENU_SURROUND_REGION:
 			_dock._surround_selection_with_region()
+		_dock.ROW_MENU_REPLACE_OBJECT:
+			_dock._open_replace_object_dialog()
 		_dock.ROW_MENU_PASTE:
 			_dock._on_paste_requested()
 		_dock.ROW_MENU_DELETE: